[OF-1586] refac: CreateAvatar Continuations (#671)
* [OF-1586] feat: Extend SignalRConnection::Invoke to be a continuation enabled method.

Allow chaining from invoke. Also involves updates to existing mocks to conform to new interface.

* [OF-1586] feat: Add SignalR continuations utility

This will be needed later. This unwraps a signalR return (from Invoke), which is always a tuple<value, exception_ptr>, and either throws the exception, or optionally forwards the result.

* [OF-1586] feat: Add internal SpaceEntity constructor

Needed later. We were doing big construction logic in SpaceEntitySystem abusing friend, this allows us to not do that, although not to remove friend because other places use it :(

* [OF-1586] feat: CreateAvatar continuations

Transform CreateAvatar to be continuations based.
The flow is

- Get the avatar ID from the server (RemoteGenerateNewAvatarId)
- Send the new avatar to the server (SendNewAvatarObjectMessage)
- Create the avatar locally (CreateNewLocalAvatar)

You'll see I've gone to some effort to prevent the possibility of leaks on error. This involves instantiating the avatar twice, (although not doing the network interaction twice, as we reuse the prior continuation result via the when_all branch) as we can't properly thread shared memory through a split continuation unless we've somewhere to put it at the end, which would have been too large a refactor. See comment in code.

Not sure about the specific continuation names, or even if I've quite got the concept communication right (especially for SendNewAvatarObjectMessage) open to change.

* [OF-1586] test: Test CreateAvatar Continuations

With the SignalRConnection mocked, we can test these chaining functions individually.
Not sure how I feel about it, it's a bit experimental. Good to explore these slight variations in testing strategy imo, a standard will settle. Still has as much coverage, or more, than testing just the top level function, and we do test that function in addition.

* [OF-1586] chore: Deal with warnings as errors

WX got added during development, which is cool. Deal with some unreferenced parameters.