[OF-1653] test: Add shared library initialization
Encapsulate the library initialization into a shared module.

The reason this is done is because we will want to use CSP in the uvu (typescript) tests in order to validate expectations, do pre-work prior to puppeteer invocation, or even just to directly do tests there when browser involvement is more headache than it's worth.

We also need to initialize CSP in the browser launched by puppeteer, so this prevents us duplicating that logic, and ensures we're on the same tenant and such.