[OF-1653] test: Create puppeteer launch framework
Include puppeteer, and create a utility function to launch a puppeteer page, and return logs and errors back to the calling function such that they can run assertions on them.

Login credentials are passed by the framework as uri arguments if provided, as they'll likely be needed in most every test.

Also port the create test user function over from primary CSP tests as we need it.