[OF-1653] test: Launch webserver to serve html tests
I tried very hard to make this work locally, but browsers (puppeteer uses chromium by default) make it extremely, impossibly difficult to get around all CORS errors using the file:// protocol, even with the 10+ security disable flags I tried to invoke it with.

As such, launch a local fileserver to serve the pages as part of the test. I'm not _super_ happy about this as I think it's rather fragile. Note how we junction the required JS that the html pages need to run into the directory we're serving.

I chose caddy simply because the web seemed to think it was the most stable and least fragile, and it is available on package managers. The ability to configure via a committed file is also nice.