[OF-1653] test: Puppeteer HTML tests (#693)
Add puppeteer based HTML tests.

This is take 2 of our prior attempts at WASM platform tests, after discovering that even a node JS executor is not good enough to properly replicate some of the gnarly emscripten issues we were seeing, we need a true browser context.

This gives the best of both worlds in a way, still a proper test framework that call call into the CSP typescript bindings from the more stable and workable node context, but also a pattern to invoke HTML browser tests to replicate browser behaviour.