[OF-1669] refac: Create IRealtimeEngine interface
Part of an ongoing modularity effort, create an IRealtimeEngine interface, containing the proposed interface for what is currently the multiplayer module.

This is mostly a port of `SpaceEntitySystem`, removing all the interfaces that are not relevent for a generic multi/single-user abstraction, as well as all the interfaces that are only used internally to what will become the RealtimeEngine module.

Changed a couple of enums to have declared types so they could be forward declared.