[OF-1669] refac: Create IRealtimeEngine interface (#700)
* [OF-1669] refac: Create IRealtimeEngine interface

Part of an ongoing modularity effort, create an IRealtimeEngine interface, containing the proposed interface for what is currently the multiplayer module.

This is mostly a port of `SpaceEntitySystem`, removing all the interfaces that are not relevent for a generic multi/single-user abstraction, as well as all the interfaces that are only used internally to what will become the RealtimeEngine module.

Changed a couple of enums to have declared types so they could be forward declared.

* fixup! [OF-1669] refac: Create IRealtimeEngine interface

Remove redundant eventbus declaration

* fixup! [OF-1669] refac: Create IRealtimeEngine interface

Stop being silly with forward enum declarations

* fixup! [OF-1669] refac: Create IRealtimeEngine interface

Use const ref because the enum is forward declared. I think this is just a generator issue, I think it's fine to use a forward declared enum like this in base c++, but could be wrong

* [OF-1669] build: Fully qualify event handler storage namespace

This prevents a conflict in the csharp if you have ASYNC declarations in a type in the `csp::common` namespace, as the List type wants to be `System.Collections.Generic`.

* fixup! [OF-1669] refac: Create IRealtimeEngine interface

Remove precondition that callback must be set. This will be determined by the concretion