[OF-1674] refac: Split continuation utils into modules
Split continuation utilities into 3 files. Common, Multiplayer (RealtimeEngine), and Systems. Systems continuations are concerned with HTTP results, multiplayer are not.

I'm quite happy with how this will let us have opinions on error handling per-module.

The largest change here is removing the concept of HTTP error codes when logging errors in a multiplayer context.

Error handling was and remains a bit confused between the two modules, I think the way forward is to split the Result heirarchy in a similar way these continuations are split. Having a Common result that is minimal, then a HTTPResult and _maybe_ a RealtimeEngineResult.

Some files still include both multiplayer and systems utils, which can't happen (outside of tests). One will be addressed in a coming commit, another is being addressed in Matts ticket.