[NT-0] build: Add libfmt
Been on the edge of adding this for a while, and this is the straw that broke the camels back.
Add fmt such that the python style formatting can be used to replace the FORMAT macros
in the multiplayer module.

Note, we must stop using the macros in any module that dosen't have singleton style access,
and we should too, macros are bad, global state is bad.