[OF-1674] refac: Have multiplayer error checking in common
This is an unfortunate, but pragmatic decision.

We have a place in SpaceSystem where we call into SpaceEntitySystem to refresh the scopes. This is probably a mistake, we want to investigate just not doing this.

However, that presents the problem of handling the errors. We need/want to call a callback on failure, which in the normal world would mean passing that callback to SpaceEntitySystem.

Double however, we don't have Results types factored that can be split nicely between the common/realtimeengine/systems trinity at the moment, so you can't do that.

To try and avoid doing that work, do this instead as a little bit of a fudge.