[OF-1672] refac: Switch to reference semantics
I was so far in CSP-generator brain that I completely forgot about references.
References should be used to communicate when a type has a non-null pointer invariant, which I'd been neglecting.
I forgot about this so hard I was considering bringing in the GSL for gsl::not_null.

Update all existing things that should be reference invariants.