[OF-1674] refac: Break cross-module dependencies in `csp::common` (#706)
* [OF-1674] refac: Extract default scheduler

This is more or less what we were doing before, just separated. It's a bit dangerous #including "async++.h" directly when we want to do this, as we need to make sure that the custom scheduler is set in the translation unit every time, or we'd accidently fall back to the threaded scheduler which would not work on WASM.

This new approach creates a easy to follow rule to just include `CSPAsyncScheduler.h` rather than `async++.h` in every case.

NOTE: You will only see files that directly include "async++" updated in this commit. Some may not even get the new include put in as they are changed to include continuation util files. The utilities get refactored later to include the scheduler, so the include propogates through there.

* [NT-0] docs: Guidance for using Local MCS

I always forget.

* [OF-1674] refac: Split continuation utils into modules

Split continuation utilities into 3 files. Common, Multiplayer (RealtimeEngine), and Systems. Systems continuations are concerned with HTTP results, multiplayer are not.

I'm quite happy with how this will let us have opinions on error handling per-module.

The largest change here is removing the concept of HTTP error codes when logging errors in a multiplayer context.

Error handling was and remains a bit confused between the two modules, I think the way forward is to split the Result heirarchy in a similar way these continuations are split. Having a Common result that is minimal, then a HTTPResult and _maybe_ a RealtimeEngineResult.

Some files still include both multiplayer and systems utils, which can't happen (outside of tests). One will be addressed in a coming commit, another is being addressed in Matts ticket.

* [OF-1674] refac: Re-namespace log types from systems->common

The logger is moving into common. Renamespace the types neccesary for that.

This does not actually move the logger, trying to keep commits smallish.

* [NT-0] build: Add libfmt

Been on the edge of adding this for a while, and this is the straw that broke the camels back.
Add fmt such that the python style formatting can be used to replace the FORMAT macros
in the multiplayer module.

Note, we must stop using the macros in any module that dosen't have singleton style access,
and we should too, macros are bad, global state is bad.

* [NT-0] feat: Add common::String formatter

Just add a formatter so we can pass common::String directly into fmt::format without having to cast to std::string.

* [OF-1674] refac: Move LogSystem to common

Move LogSystem from systems -> common.
Update the systems manager to instantiate it from that namespace

* [OF-1674] refac: Break continuation LogSystem dependencies

Break the continuation dependency on SystemsManager (in continuation util functions in the common and realtimeEngine modules), and update files that use them to provide the LogSystem via injection

* [OF-1674] refac: Move enums neccesary for common continuations to shared location

We'll need to do stuff like this to break dependencies, although we should resist things going into
common that don't need to be.

* [OF-1674] refac: Migrate refresh scopes to SpaceEntitySystem

Really where it should have been the whole time, since it exclusively uses multiplayer based types.

... I'd like to not have this step, feels mega redundant. Need time to investigate.

* [OF-1674] docs: Add new cross module dependency to method listing

Keep the plan up to date.

* [OF-1674] refac: Have multiplayer error checking in common

This is an unfortunate, but pragmatic decision.

We have a place in SpaceSystem where we call into SpaceEntitySystem to refresh the scopes. This is probably a mistake, we want to investigate just not doing this.

However, that presents the problem of handling the errors. We need/want to call a callback on failure, which in the normal world would mean passing that callback to SpaceEntitySystem.

Double however, we don't have Results types factored that can be split nicely between the common/realtimeengine/systems trinity at the moment, so you can't do that.

To try and avoid doing that work, do this instead as a little bit of a fudge.

* [OF-1674] refac: Remove conflicting dependency

The whole point of the prior commit :P