[OF-1672] refac: Inject logsystem into entity script binding
What is unique about this one is that it replaces the `Log` script function away from a function pointer and towards a capturing functor object, (in order to capture the logsystem).

... starting to wonder if we need to document lifetime requirements for our system. Before, the idea of having a CSP that didn't start and end all at once as a massive block was just unimaginable, but now... I wonder if we might hit problems with the logsystem not being in-memory at some points if people construct and destruct CSP in out-of-order ways.

Back on point though, I think this is tested well enough ,we have several tests that test the log function called from a script.