[OF-1672] refac: Undo reference semantics, revert to unsafe pointers
Because of the wrapper generator, revert the storage that has to be pointers
back to pointers.

Not rebasing this change away because the story of how it should be done up until this point is clear, this is an unfortunate compromise with reality that irks me, and a direct technical reason behind why much of our instability manifests.