[OF-1672] refac!: Remove SystemsManager dependency based on getting the LogSystem from multiplayer. (#716)
Remove all dependencies that deliver `LogSystem` to the multiplayer module.

Some `Debug/Logging` includes remain due to the annoying profiling macros being contained there too, we'll have to either disable that functionality for the realtimeengine module or do the work to port it.

- All components gain an optional `LogSystem` member via injection.
- `SpaceEntities` gain an optional `LogSystem` member via injection
- `SystemBase`, and thus all systems, gain a mandatory `LogSystem` member via injection
- Various other elements of the codebase gain a `LogSystem` member via injection, where there is no public constructors in play, this is expressed when possible via a non-nullable reference, otherwise the non-nullability is only weakly captured via code comment
- Due to unravelling this ball of mud, some transient includes are now manifest and required to be directly included, they are placed with a note that the dependency needs breaking.
- Coincidentally fixed us not deleting the conversation system

No new tests are created, as this is a pure refactor. The most risky change is the change to `ScriptSystem::Log`, as it becomes a capturing lambda now so as not to abuse static data. However, this is quite well tested so I'm not worried.


Breaking changes.

Several types now take a `LogSystem` in their construction, which can be fetched from `SystemsManager`.
- All Components
- EntityScript
- SpaceEntity

For the above types, this is an optional argument, you may pass `nullptr` to these to disable their logging functionality.

Breaking changes.

Several types now take a `LogSystem` in their construction, which can be fetched from `SystemsManager`.
- All Components
- EntityScript
- SpaceEntity

For the above types, this is an optional argument, you may pass `nullptr` to these to disable their logging functionality.