[OF-1672] refac: Create IJSScriptRunner interface
Several types in the multiplayer module access the ScriptSystem, which is remaining in Systems.

This is the starting of a refactor that will be messy, as we all know scripting is a mess and isn't factored properly. Using this interface we can at least only talk in terms of functionality that is required for the RealtimeEngine module.