[OF-1672] refac!: Move LoginState to common
Move the loginstate to common.
This type is used in a few flows, like avatar creation as well as in the WebClient (which is in common for now, although I'm not a massive fan of that.)

Seems like a good thing to be able to pass around as an object between modules.