[OF-1672] refac!: Use LoginState to break SystemManager dependencies in SpaceEntitySystem
Break the dependency, provide the common LoginState object to all the calls that need it, ie, the CreateAvatar calls.

This is not _neccesarily_ the best api ... it may be better for the RealtimeEngine to be stateful and take a LoginState on construction so you don't have to provide this. Although that _is_ worse from a decoupling perspective and puts onus on the user to update the login state when it changes ... although it means you don't have to pass it all the time.