[OF-1672] refac!: Break all ScriptSystem dependencies in RealtimeEngine module (#724)
* [OF-1672] refac: Create IJSScriptRunner interface

Several types in the multiplayer module access the ScriptSystem, which is remaining in Systems.

This is the starting of a refactor that will be messy, as we all know scripting is a mess and isn't factored properly. Using this interface we can at least only talk in terms of functionality that is required for the RealtimeEngine module.

* [OF-1672] refac: Extract existing IScriptBinding interface

This is a handy pre-existing interface that we can yank to common in order to break some dependencies between the Binding (in RealtimeEngine) and the runtime and scriptSystem (in Systems)

* [OF-1672] refac!: Break dependency with ClientElectionManager

This involved injecting an IJSRunner into SpaceEntitySystem, which is the main point of dependency injection for the new module, and demonstrates the primary strategy for this dependency break.

* [OF-1672] refac: Break dependency with ClientProxy

Injected the interface from ClientElectionManager in the prior commit.

* [OF-1672] refac: Make Scriptsystem conform to interface

So you can pass scriptsystem anywhere you pass the IJSRunner interface. We've already been doing this up to this point, just get it into the changeset.

* [OF-1672] refac: Break binding dependencies

This also makes the script runner / system a non optional parameter to the binding, which is possible because this is not a public type that has to go through wrapper gen.

* [OF-1672] refac!: Break dependency with SpaceEntity

Standard injection from SpaceEntitySystem

* [OF-1672] refac!: Break dependency with EntityScript

Require an injection into entityScript. Standard dependency injection.

This is a purely public type, will definatley require changes in clients, probably ones that seem weird and unnecesary from their perspective.

* [OF-1672] refac: Make IJSScriptRunner a fake interface

Extremely annoying. Because this is a public type, our C# and typescript interfaces cannot coexist in a way that allows us to have a true interface (with pure virtual methods) that is passed across public API.

Do this hack for now.

* [OF-1672] docs: Fix spelling

* [OF-1672] fix: Move pragma to top

* [OF-1672] fix: Remove warning pragmas + returns

---------

Co-authored-by: MAG-mv <109593178+MAG-mv@users.noreply.github.com>