[OF-1702] refac!: Rewrite EventBus
Rewrite EventBus to only manage callbacks, removing the concept of systems from it entirely.
Also change the deserialization such that EventBus manages it directly, removing the need for systems to know about deserializers.
Introduce a new EventData object that is used to pass around ReplicatedValues, and can be cast to more concrete data by types that have registered for special event handling.