[OF-1702] refac: Update Conversation types to use EventData from Params
This commit is split in this weird way just because the deserialization changed slightly for ConversationHelpers, in that we provide more specific arguments to `MessageInfoToReplicatedValueArray`. This is done to avoid having to rewrite a bunch of utility functions. We have the information we need to perform the logic anyhow, so this is easiest.

The rest of the Conversation refactors are in this commit just as an organisational thing, the rest of the updating from Params->EventData comes in the next commit.