[OF-1702] refac: Hand-roll polymorphism in SequenceEvent
One of the worst hacks I've done, although there are reasons beyond high code entropy. Specifically, the prior polymorphism based hack cannot work due to lack of RTTI information on WASM (whether this is a good idea or not is up for debate)

What we do here is something like you might find in a C codebase, use a pointer (or in this case an optional ... because wrapper generator) to an additional bundle of data that may or may not be on the type (think subclass), and check that for nullness/emptyness. If we moved this to the top of the type and called it `typeinfo` you'd be halfway to a C++ polymorphic data layout.

I very much dislike this, mostly because the `HotspotData` dosen't make intuitive sense as data related to hotspots. I have no idea why these additional data members are special and neccesary for hotspot workflows. It's bad, it's going to be confusing, I can't see another way of doing it without getting away from refactors and into workflow design.