[OF-1672] refac!: Break almost all remaining Systems dependencies in RealtimeEngine module (#726)
Other than the conversation system, which needs an eye from Matt, this breaks all other SystemManager/System dependencies in the multiplayer namespace.

This is mostly normal dependency breaking stuff. The main thing is LoginState moving to common. This is a good type to be in common as knowing the state of a user login session is something all modules can have a legitimate interest in.

Other than refactors, this changes the behaviour of the HotspotSpaceComponent when used in a hotspot sequence. See breaking change + note below.

Breaking Changes

- SpaceEntitySystem now takes an EventBus as a constructor argument
- csp::systems::LoginState now moved to csp::common::LoginState, now found in Common/LoginState.h
- csp::systems::ELoginState moved to csp::common::ELoginState, now inside Common/SharedEnums.h
- SpaceEntitySystem::CreateAvatar now takes a LoginState as a parameter (which can be fetched from - UserSystem.GetLoginState)
- HotspotSpaceComponent no longer automatically deregisters itself from any sequences it is in upon deletion. Clients should call HotspotSequenceSystem.RemoveItemFromGroups(HotspotComponent.GetUniqueComponentId()) manually if they wish to deregister a deleted hotspot. See the test in this PR for an example.

* [OF-1672] refac: Break EventBus dependency in ClientProxy

Inject it as usual.

This is a more controversial one, as the SpaceEntitySystem actually has access to the EventBus already via the MultiplayerSystem. However, I felt this relied too much on internal knowledge of how SystemsManager is currently creating types ... although only very loosely, could change this easily.

* [OF-1672] refac: Break Systemsmanager dependency in MultiplayerConnection

Easy break, just needed it for SpaceEntitySystem which is available in the module anyhow.

* [OF-1672] refac: Break unused UserSystem dependency in SignalRClient

LoginState just wasn't used at all.

* [OF-1672] refac!: Move LoginState to common

Move the loginstate to common.
This type is used in a few flows, like avatar creation as well as in the WebClient (which is in common for now, although I'm not a massive fan of that.)

Seems like a good thing to be able to pass around as an object between modules.

* [OF-1672] refac!: Use LoginState to break SystemManager dependencies in SpaceEntitySystem

Break the dependency, provide the common LoginState object to all the calls that need it, ie, the CreateAvatar calls.

This is not _neccesarily_ the best api ... it may be better for the RealtimeEngine to be stateful and take a LoginState on construction so you don't have to provide this. Although that _is_ worse from a decoupling perspective and puts onus on the user to update the login state when it changes ... although it means you don't have to pass it all the time.

* [OF-1672] refac: Break SpaceEntity dependency on global state

Inject the bool neccesary as a parameter.

* [OF-1672] refac!: Remove functionality that has HotspotSpaceComponent deregister itself.

It's a little weird that it does this in general.

Clients will now have to remove the hotspot space component from the sequence system groups themselves when they remove it. This is to break the dependency on SystemManager.

* fixup! [OF-1672] refac!: Move LoginState to common

Remove copy/paste error comment