[OF-1702] refac!: Rewrite EventBus to break dependencies and reduce design confusion (#731)
Rewrite the EventBus interfaces to no longer be aware of Systems (in particular SystemBase), and favour only callback registration., in order to break dependencies for modularisation.

We solve the problem about not being able to identify callbacks by introducing a NetworkEventRegistration type, which asks for an identifier ID which is used as the discriminator. By doing this, we also enable more powerful api usage, such as clients registering multiple times for the same event, or grouping/managing events under specific IDs. This isn't the point of the change, but it's a nice side-effect nonetheless. Some API is added to support this.

Finally, we change the way the deserialization works such that the EventBus is aware of how to deserialize the events that require custom deserialization. Doing this allows us to pass a polymorphic EventData object as the exchange type via the callbacks. This wraps ReplicatedValue, which is moved to common and is now usable as an interop type. We also make it structurally clear that the custom-deserialization types are special cases.

Also rewrote the event bus test suite from scratch, as its simpler interfaces can now be served by a simpler test suite.