[NT-0] ci: Update the PR Guidance workflow to support forks (#747)
When a pull request is created from a fork, GitHub Actions workflows run in a more restricted context for security reasons. The 'pull_request_target' event is used to give the workflow the appropriate write permissions and runs the workflow in the context of the base repository (the one the PR is targeting).