[OF-1670] refac!: RealtimeEngine connection flow, memory deletion, proxy lifetimes.
This is weird commit.

To get this working, I needed to do red/green refactors as the scale of this change is so large.
This commit moves to the "new" connection flow, which is mostly concerned with calling IRealtimeEngine::FetchAllEntitiesAndPopulateBuffers in the EnterSpace call.

The underlying idea here is that IRealtimeEngine will be provided to EnterSpace, and owned by the user.
However, I couldn't do that all at once, (believe me ... I tried...), so this change does a bunch of hacks to keep the
current SpaceEntitySystem lifetime the same as it was before, whilst using the connection interfaces
for the new thing.

BIG DEAL: Note how we're accounting for how RealtimeEngine might be null when recieving SignalR messages. This was the motivation behind the changes to the bindings. Previously, SpaceEntitySystem lived forever, munging its state to give itself psuedo "connected" and "disconnected" states (whilst also leaving massive room for a whole host of re-enter space bugs.)
The use of weak_ptr allows us to know if the SpaceEntitySystem has been deleted, (which we do in ExitSpace to simulate clients doing this), allowing us to never be capable of forwarding signalR messages to a deleted IRealtimeEngine, remember even if you have deleted an object, if you have a lambda registered as a callback it will still be called, even if it uses the `this` ptr which is zeroed (but not nulled!). Without checking for this, this class of error would certainly occur during ExitSpace.

In the new future (without any owned pointer semantics over the wrapper gen), we'll have to say to clients that they are required to keep their IRealtimeEngine memory alive until ExitSpace has completed.