[OF-1670] refac!: IRealtimeEngine interface fake interface
Incredibly annoyingly, when you try to actually _use_ an interface, the wrapper gen can't handle it, so we have to do this and be fake. It's very frustrating.

This also changes the arguments to CreateAvatar and CreateEntity slightly, in service to refactoring later. (Removing a bunch of the internal methods). Think about this, this is nicer code and provides more power to clients, but were the defaults better? Easy to just hard code the defaults if so ... (although again, if the wrapper gen were better we could express this cleanly in the one interface without applying additional client burden via overloads)