[OF-1670] refac!: MakeSpaceEntitySystem a RealtimeEngine
Inherit from the IRealtimeEngine interface.

As IRealtimeEngine is designed around SpaceEntitySystem, this is mostly header file work, the implementations and most of the method names remain the same.

Have tried to add at least some documentation to any methods that were touched, even ones that didn't change.

... I'd love to clean this up completely, have had to resist, feeling like I understand it well enough at this point after learning through all the errors in this refactor to do it well.

NOTE: This commit (and all the others) can't help but roll up some of the other refactoring work. I'm trying to keep them as seperated as I can but you can imagine it's tricky. Sorry for the bother. If you're worried that a method has dissapeared and can't spot where in the changeset/other commits it is, shoot a question and i'll either find it, or realize I've made a mistake.