[OF-1670] refac: Remove initial-entity-fetch transitional shim
To get the tests to stay green during refactor, we has a shim in enterspace to set it from static state on the SpaceSystem. This finishes the transition and gets it from the RealtimeEngine itself.

This will probably seem like a strange commit in isolation, but remember there were hundreds of tests needing changed along with this, i was trying to make small steps to validate my assumptions every step of the way.