[OF-1670] test: Procedural updates to tests
Update the interfaces in the tests to the new one.

Majority thrust of this is that each test that enters a space will now be making a realtime engine, and those realtime engines will be forced to set an initial entity fetch callback.

There's some other stuff, like setting the optional parameter for parentID