[OF-1670] test: Mock injection infrastructure for SignalRConnection
I didn't really want to do this quite yet, it's more to do with the work we've planned for making the MultiplayerConnection a point of dynamic dispatch.

However, plans all hang together and it's hard to do bits of them in isolation, and the need to inject mocks forced my hand on this. (Testable code == well designed code afterall)

Make the ISignalRConnection an invariant on MultiplayerConnection. This is weird because MultiplayerConnection is still static and lives on the SystemsManager, (for now), which was _fine_ because connect hooked up a brand new SignalRConnection every login before.

We don't do that anymore, because we want to test more thoroughly by injecting mocks, and also setting an object member to null/not-null based on execution flow is fragile. The small hack we do to make this work is guard against double-setting the `on` method, which will happen if you attempt to login for a second time in the same CSP session. (The `on` method throws on double sets ... i might actually come back and just make this exception handling behaviour ... that is probably better than using state. )

The initialisation flow is changed to service the possible (non exported) injection of a mock signalR object. This pattern can be expanded to inject other mock objects, like scripting systems, or REST systems, later.

Also means we can stop guarding against connection being null and getting weird states in our branch tree, which is nice. Only a small distance away from needing to track `Connected` as well, and just have a `MultiplayerConnection` object be valid by definition.