[OF-1670] fix: Fatal error path for exiting space
Null the realtime engine when exiting a space.

I have defined this path as a fatal error, but nulling out the engine does seem like the sensible thing to do in any case.

Went to write happy path and sad path tests for this, but found that whilst mocking the CreateSpace/EnterSpace flow is possible now, it's quite a lot of work, we should get round to doing it but I don't think I can feasibly roll it in.