[OF-1732] fix!: Shorthand signiature wrapper gen bug workaround
Workaround the wrapper generator bug with shorthand signature generation.

There is an issue where if the return type of the accepting function is different from the return type of the registering callback (at least in cases like this where we're registering for later behaviour, not making a task/promise sort of construct), then the shorthand type sig used for binding callbacks will generate wrong, using the return type of the enclosing function.

Coincidently, we'd never done this before.

Wrapper generator bug still exists, this just avoids it.