OF-1714 fix: Ensure internal headers compile with warnings (#759)
We currently classify all includes as external includes which mean they are getting compiled without warning.

Updated the premake file to remove src and include directories from 'externalincludedirs' and to add them to 'includedirs' instead.

This resulted in a number of wErrors which needed to be resolved. I have endeavoured to group these changes by error type as much as possible. This is reflected in the commits.