[OF-1670] refac!: Online Realtime Engine (#755)
Converts SpaceEntitySystem to OnlineRealtimeEngine, providing an interface for clients to pass arbitrary IRealtimeEngine instances to EnterSpace, such that CSP can branch on different engine behaviour. The most immediate desire for this being the support of an OfflineRealtimeEngine which is coming soon, but also opens up lots of alternate behaviour, such as a more lean ServerSideScriptRunnerRealtimeEngine or even a MockRealtimeEngine, etc. I'm quite excited about how this interface point makes experimental implementations of radically new communication/entity management techniques possible.

This has many implications for our memory management during login and space entry. In general, you will see a removal of manual state management in objects, and an addition of constructor arguments being used to establish invariants. This all takes into account the future direction of the EnterSpace flow, and is a solid step towards the Summer Party Vision, as well as being bettter in general.

An important factor is the seperation of SignalR binding and dispatch, you'll see this in early commits. MultiplayerConnection is now explicitly bound during Connect, and maintains those bindings for its entire lifetime. Those bindings are only actually dispatched to a RealtimeEngine if one has been set, which is something that is invariant to the temporal scope of entering a space (as defined by the scope between the instruction when EnterSpace is first called, and the receiving instruction of the ExitSpace callback)
So long as this invariant holds, it is not possible to dispatch an off-thread network event to null memory. Remember, whilst RealtimeEngine has been extracted the way we want it, MultiplayerConnection as yet has not, so a shim is neccesary.

There are a ton of not-great decisions made here due to the wrapper generator, it's really been pushed to its limit during this work. (See this for a listing of some of the issues that occurred during this work.)

The most major of these compromises is the need for clients to manually ensure RealtimeEngine lifetime for the scope of being in a space. Ideally this would be expressed in the interface via a shared_ptr/weak_ptr link, or a different memory management strategy, but we're not there yet.