[OF-1689] feat: Added public interface for OfflineRealtimeEngine
Just the header file for OfflineRealtimeEngine, implementing the interface from IRealtimeEngine. Docs are similar, but not the same as the OnlineRealtimeEngine, as this verison is synchronous and does not use SignalR. There is still a TODO for FetchAllEntitiesAndPopulateBuffers, as it currently doesn't do anything, as entities will be populated in the constuctor for the time being. This will likely change when the Offline changes have been merged