[OF-1689] refac: Refactored uuid generator by seperating the random integer creation into a seperate function
This new function will be used for generating the id of SpaceEntities in the OfflineRealtimeEngine. This wasn't needed for the oniline version, as MCS generate ids for us. If we ever want to create an OnlineSpace from an offline version, I am not sure if this id will be compatible. This isn't a concern for the time being