[OF-1689] refac: Updated CSPSceneDescription to work with OfflineRealtimeEngine
Both classes needed to inject each other in their constructors, so the entitiy creation within the CSPSceneDescription is now done in CSPSceneDescription::CreateEntities. There is likely a better way of doing this. However, this is a comprimise so I didn't have to change the tests too much. Clients wont be interaction with this function, so I'm not concerned with potentially changing this in the future