[OF-1689] test: Added OfflineRealtimeEngine tests
I did TDD, so tests were done before implementation

When The OnlineRealtimeEngine changes have been merged, we will want to add additional tests for the EnterSpace flow, as this can't be done yet. The OfflineRealtimeEngine has only been tested in isolation. We also need to chack that the ParentTest is working, as this is currently crashing due to the SpaceEntity having a depdendency on SpaceEntitySystem instead of IRealtimeEngine