[OF-1689] feat: Added shared RealtimeEngine functionality
As mentioned in the comments, there is likely a much better way of handling this, as Online should be an extension of Offline. For now, these shared functions have been created to reduce code duplication between both implementations. I have also had to make some functions in SpaceEntity and ComponentBase public, as we cannt rely on friendship anymore.