[OF-1689] feat: Main OfflineRealtimeEngine functionality
The implementation is actually pretty simple, as a lot of this functionality is shared with the Online version. We also dont need to deal with any threads, so we dont need to defer updates. We are still using ProcessPendingEntityOperations, but dont for EntityUpdates, such as property and component changes, as this functionality is built into the Entity itself. I also haven't figured out what to do with the FetchAllEntitiesAndPopulateBuffers, as entity creation is currently done in the constuctor. I couldn't test the full EnterSpace flow at this point, as it relies on the Online refactor work